perf(shared): move `mergeDocConfig` to node utils
