+++ packages/core/src/index.ts
@@ -1,2 +1,3 @@
 export * from './node';
 export * from '@rspress/shared';
+export { mergeDocConfig } from '@rspress/shared/node-utils';
