+++ packages/runtime/package.json
@@ -54,7 +54,6 @@
     "./src/theme-default/styles/index.ts"
   ],
   "files": [
-    "bin",
     "dist",
     "src",
     "runtime.ts",
