+++ packages/shared/package.json
@@ -56,18 +56,15 @@
     "execa": "5.1.1",
     "fs-extra": "11.2.0",
     "gray-matter": "4.0.3",
+    "lodash-es": "^4.17.21",
     "unified": "^10.1.2"
   },
   "devDependencies": {
     "@types/fs-extra": "11.0.4",
-    "@types/interpret": "^1.1.3",
     "@types/jest": "~29.5.14",
-    "@types/lodash": "^4.17.13",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^18.11.17",
     "@types/react": "^18.3.12",
-    "@types/rechoir": "^0.6.4",
-    "lodash-es": "^4.17.21",
     "medium-zoom": "1.1.0",
     "rimraf": "^3.0.2",
     "typescript": "^5.5.3"
@@ -80,7 +77,6 @@
   },
   "license": "MIT",
   "files": [
-    "bin",
     "dist"
   ]
 }
