+++ packages/shared/src/node-utils.ts
@@ -1,1 +1,2 @@
 export * from './node-utils/loadFrontMatter';
+export { mergeDocConfig } from './node-utils/merge';
