+++ packages/shared/src/runtime-utils/index.ts
@@ -1,6 +1,3 @@
-import { castArray, isArray, isUndefined, mergeWith } from 'lodash-es';
-import type { UserConfig } from '@/types';
-
 export const QUERY_REGEXP = /\?.*$/s;
 export const HASH_REGEXP = /#.*$/s;
 export const MDX_REGEXP = /\.mdx?$/;
@@ -213,14 +210,6 @@ export function replaceVersion(
   );
 }
 
-export const omit = (obj: Record<string, unknown>, keys: string[]) => {
-  const ret = { ...obj };
-  for (const key of keys) {
-    delete ret[key];
-  }
-  return ret;
-};
-
 export const parseUrl = (
   url: string,
 ): {
@@ -288,21 +277,3 @@ export function removeBase(url: string, base: string) {
     '',
   );
 }
-
-export function withoutHash(url: string) {
-  return url.split('#')[0];
-}
-
-export const mergeDocConfig = (...configs: UserConfig[]): UserConfig =>
-  mergeWith({}, ...configs, (target: UserConfig, source: UserConfig) => {
-    const pair = [target, source];
-    if (pair.some(isUndefined)) {
-      // fallback to lodash default merge behavior
-      return undefined;
-    }
-    if (pair.some(isArray)) {
-      return [...castArray(target), ...castArray(source)];
-    }
-    // fallback to lodash default merge behavior
-    return undefined;
-  });
